@@ -1,4 +1,4 @@
-name: note_berkat
+name: versus
 description: A new Flutter project.
 
 # The following defines the version and build number for your application.
@@ -11,38 +11,47 @@ description: A new Flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.0.6+7
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  rxdart: ^0.22.0
-  toast: ^0.1.5
-  shared_preferences: ^0.5.6+3
-  lottie: ^0.3.3
-  dio: ^3.0.9
-  pretty_dio_logger: ^1.1.1
-  provider: ^4.1.0
-  firebase_core: ^0.4.0+9
-  firebase_auth: ^0.14.0+5
-  firebase_database: ^3.0.7
-  firebase_messaging: ^4.0.0+1
-  flutter_staggered_grid_view: ^0.3.0
-  intl: ^0.16.1
-  after_layout: ^1.0.7+2
-  flutter_bloc: ^4.0.0
-  flutter_local_notifications: ^1.4.3
+  rxdart: ^0.27.7
+  toast: ^0.3.0
+  shared_preferences: ^2.2.2
+  lottie: ^2.3.2
+  dio: ^5.4.0
+  pretty_dio_logger: ^1.3.1
+  provider: ^6.1.1
+  #package_info_plus: ^4.2.0
+  flutter_masked_text2: ^0.9.1
+  intl: ^0.18.1
+  after_layout: ^1.2.0
+  currency_text_input_formatter: ^2.1.10
+  number_pagination: ^0.0.3+2
+  floating_action_bubble: ^1.1.4
+  syncfusion_flutter_datepicker: ^22.2.12
+  image_picker: ^1.0.4
+  characters: ^1.2.1
+  url_launcher: ^6.1.11
+  file_picker: ^5.3.1
+  shorebird_code_push: ^1.1.3
+  collection: ^1.15.0-nullsafety.4
+  badges: ^3.1.2
+  restart_app: ^1.2.1
+  flutter_linkify: ^6.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
 
 
 # For information on the generic Dart part of this file, see the
@@ -59,6 +68,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - images/
+    - shorebird.yaml
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
